+++ kubernetes-model-generator/generateModel.sh
@@ -50,6 +50,7 @@ declare -a modules=(
     "openshift-model-monitoring"
     "openshift-model-tuned"
     "openshift-model-whereabouts"
+    "openshift-model-storageversionmigrator"
 )
 generateAll() {
   for module in ${modules[*]}